requirements.txt
reportlab
flask_sqlalchemy
flask_migrate
flask_upload
pymysql
faker
